requirements.txt
asgiref==3.7.2 ; python_version >= "3.11" and python_version < "4.0"
cfgv==3.4.0 ; python_version >= "3.11" and python_version < "4.0"
distlib==0.3.8 ; python_version >= "3.11" and python_version < "4.0"
django-cors-headers==4.3.1 ; python_version >= "3.11" and python_version < "4.0"
django==5.0 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework-simplejwt==5.3.1 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.11" and python_version < "4.0"
filelock==3.13.1 ; python_version >= "3.11" and python_version < "4.0"
identify==2.5.33 ; python_version >= "3.11" and python_version < "4.0"
nodeenv==1.8.0 ; python_version >= "3.11" and python_version < "4.0"
platformdirs==4.1.0 ; python_version >= "3.11" and python_version < "4.0"
pre-commit==3.6.0 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.11" and python_version < "4.0"
pyjwt==2.8.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
setuptools==69.0.3 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
virtualenv==20.25.0 ; python_version >= "3.11" and python_version < "4.0"
